wip: add cyrpto core
